+++ src/app/(auth)/signup/page.tsx
@@ -32,6 +32,7 @@ import {
 	X,
 } from "lucide-react";
 import { toast } from "react-toastify";
+import Logo from "@/components/common/logo";
 
 export default function RegisterPage() {
 	const [formState, setFormState] = useState({
@@ -53,7 +54,7 @@ export default function RegisterPage() {
 	const [register, { isLoading, isError, error }] = useRegisterMutation();
 	const router = useRouter();
 	const { isAuthenticated, user, loading } = useSelector(
-		(state: RootState) => state.auth
+		(state: RootState) => state.auth,
 	);
 	const pathname = usePathname();
 	const searchParams = useSearchParams();
@@ -99,7 +100,7 @@ export default function RegisterPage() {
 			return;
 		}
 		if (passwordStrength < 3) {
-			toast.warn ("Must contain 8 characters, 1 number, 1 symbol")
+			toast.warn("Must contain 8 characters, 1 number, 1 symbol");
 			return;
 		}
 
@@ -164,17 +165,7 @@ export default function RegisterPage() {
 			<div className="w-full max-w-md">
 				{/* Logo/Brand */}
 				<div className="text-center mb-8">
-					<Link
-						href="/"
-						className="inline-flex items-center space-x-2 mb-4"
-					>
-						<div className="flex items-center justify-center w-10 h-10 bg-linear-to-br from-blue-600 to-purple-600 rounded-lg">
-							<SquarePen className="h-5 w-5 text-white" />
-						</div>
-						<span className="font-bold text-2xl bg-linear-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent">
-							MNHPSurvey
-						</span>
-					</Link>
+					<Logo/>
 					<p className="text-muted-foreground">Create your account</p>
 				</div>
 
@@ -443,7 +434,7 @@ export default function RegisterPage() {
 							{/* Submit Button */}
 							<Button
 								type="submit"
-								className="w-full bg-linear-to-r from-blue-600 to-purple-600 hover:from-blue-700 hover:to-purple-700 text-white transition-all shadow-sm hover:shadow-md"
+								className="w-full bg-linear-to-r from-gradient-l to-gradient-r  hover:from-gradient-l/50 hover:to-gradient-r/50 text-white transition-all shadow-sm hover:shadow-md"
 								disabled={isLoading}
 								size="lg"
 							>