+++ src/app/(landingpage)/layout.tsx
@@ -1,4 +1,5 @@
 import { Navbar } from "@/components/common/navigation";
+import Footer from "./_components/Footer";
 
 // frontend/src/app/(landingpage)/layout.tsx
 export default function LandingPageLayout({
@@ -12,6 +13,7 @@ export default function LandingPageLayout({
 			<main className="mx-auto w-full max-w-7xl">
 				{children}
 			</main>
+			<Footer />
 		</>
 	);
 }